requirements.txt
streamlit
numpy
pillow
opencv-python

